refactor Pharos::Phase, Pharos::Phases
