+++ lib/pharos/phases/configure_network.rb
@@ -1,10 +1,8 @@
 # frozen_string_literal: true
 
-require_relative 'base'
-
 module Pharos
   module Phases
-    class ConfigureNetwork < Base
+    class ConfigureNetwork < Pharos::Phase
       PHASE_TITLE = "Configure network"
 
       WEAVE_VERSION = '2.2.1'
